함수를 명령으로 변경 #1
