+++ src/main/java/me/whiteship/refactoring/_03_long_function/_10_replace_function_with_command/StudyDashboard.java
@@ -5,10 +5,7 @@ import org.kohsuke.github.GHIssueComment;
 import org.kohsuke.github.GHRepository;
 import org.kohsuke.github.GitHub;
 
-import java.io.FileWriter;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.Comparator;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
@@ -47,7 +44,7 @@ public class StudyDashboard {
         ExecutorService service = Executors.newFixedThreadPool(8);
         CountDownLatch latch = new CountDownLatch(totalNumberOfEvents);
 
-        for (int index = 1 ; index <= totalNumberOfEvents ; index++) {
+        for (int index = 1; index <= totalNumberOfEvents; index++) {
             int eventId = index;
             service.execute(new Runnable() {
                 @Override
@@ -72,55 +69,6 @@ public class StudyDashboard {
         latch.await();
         service.shutdown();
 
-        try (FileWriter fileWriter = new FileWriter("participants.md");
-            PrintWriter writer = new PrintWriter(fileWriter)) {
-            participants.sort(Comparator.comparing(Participant::username));
-
-            writer.print(header(participants.size()));
-
-            participants.forEach(p -> {
-                String markdownForHomework = getMarkdownForParticipant(p);
-                writer.print(markdownForHomework);
-            });
-        }
+        new StudyPrinter(this.totalNumberOfEvents, participants).execute();
     }
-
-    private String getMarkdownForParticipant(Participant p) {
-        return String.format("| %s %s | %.2f%% |\n", p.username(), checkMark(p, this.totalNumberOfEvents),
-                p.getRate(this.totalNumberOfEvents));
-    }
-
-    /**
-     * | 참여자 (420) | 1주차 | 2주차 | 3주차 | 참석율 |
-     * | --- | --- | --- | --- | --- |
-     */
-    private String header(int totalNumberOfParticipants) {
-        StringBuilder header = new StringBuilder(String.format("| 참여자 (%d) |", totalNumberOfParticipants));
-
-        for (int index = 1; index <= this.totalNumberOfEvents; index++) {
-            header.append(String.format(" %d주차 |", index));
-        }
-        header.append(" 참석율 |\n");
-
-        header.append("| --- ".repeat(Math.max(0, this.totalNumberOfEvents + 2)));
-        header.append("|\n");
-
-        return header.toString();
-    }
-
-    /**
-     * |:white_check_mark:|:white_check_mark:|:white_check_mark:|:x:|
-     */
-    private String checkMark(Participant p, int totalEvents) {
-        StringBuilder line = new StringBuilder();
-        for (int i = 1 ; i <= totalEvents ; i++) {
-            if(p.homework().containsKey(i) && p.homework().get(i)) {
-                line.append("|:white_check_mark:");
-            } else {
-                line.append("|:x:");
-            }
-        }
-        return line.toString();
-    }
-
 }
